pyproject.toml
[tool.poetry]
name = "zona-igr"
version = "0.1.0"
description = ""
authors = ["krasniy-1889 <neytirri@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
python-dotenv = "^1.0.0"
argon2-cffi = "^23.1.0"
psycopg2-binary = "^2.9.9"
django-extensions = "^3.2.3"
django-redis-sessions = "^0.6.2"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
pillow = "^10.1.0"
django-debug-toolbar = "^4.2.0"
django-tree-queries = "^0.16.1"
loguru = "^0.7.2"
cuid = "^0.4"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.0"
pymysql = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
